Clean up; hack for player memory corruption; sound fx for bomb, damage

+++ level.cpp
@@ -117,9 +117,9 @@ byte getTile(unsigned int posX, unsigned int posY) {
   );
 }
 
-void drawLevel(int posX, int posY) {
-  for (int i = 0; i < BOARD_DIM; i++) {
-    for (int j = 0; j < BOARD_DIM; j++) {
+void drawLevel(byte posX, byte posY) {
+  for (byte i = 0; i < BOARD_DIM; i++) {
+    for (byte j = 0; j < BOARD_DIM; j++) {
       int wx = i * 16 + CAM_X_OFFSET + 16 - posX;
       int wy = j * 16 + CAM_Y_OFFSET + 16 - posY;
       if (wx < 0 || wx > WIDTH + 16 || wy < 0 || wy > HEIGHT + 16) continue;
